refactor: use u128 in emit, and change nullifier

+++ contracts/precompiles/src/IEmit.sol
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: UNLICENSED
+// SPDX-License: UNLICENSED
 pragma solidity ^0.8.30;
 
 /// @title IEmit
@@ -8,7 +8,7 @@ interface IEmit {
     /// Burn native COEN into a private note. The commitment is derived from
     /// the runtime chain ID, `noteSn`, and the caller-supplied value; the note
     /// itself (owner, spend key) is chosen off-chain and proven later at mint.
-    /// `msg.value` must be a positive amount fitting `uint64` (native base
+    /// `msg.value` must be a positive amount fitting `uint128` (native base
     /// units, 1:1 with circuit units). Initializes the tree on the first call.
     function burn(bytes32 noteSn) external payable;
 
@@ -18,15 +18,16 @@ interface IEmit {
     /// deterministic change commitment. The caller must be `noteOwner`; the
     /// embedded proof statement must equal the explicit calldata fields, and
     /// `chainId` must equal the runtime chain ID. `proof` is the combined
-    /// UltraHonkKeccak wire for the frozen Emit mint circuit, version 1.0.0,
-    /// capped at 16,384 bytes.
+    /// UltraHonkKeccak wire for the frozen Emit mint circuit
+    /// (`outbe.emit.mint`, version 1.3.0), enforced at its exact frozen
+    /// length.
     function mint(
         address payoutRecipient,
         uint64 chainId,
         bytes32 root,
         bytes32 nullifier,
         address noteOwner,
-        uint64 mintUnits,
+        uint128 mintUnits,
         bytes32 changeCommitment,
         bytes calldata proof
     ) external;
@@ -37,7 +38,7 @@ interface IEmit {
     /// @param rootAfter Tree root after the append.
     /// @param noteAmount Burned public amount; `0` is the sentinel for a
     ///        partial mint's change note, whose remaining value is private.
-    event NewNote(bytes32 indexed commitment, uint32 leafIndex, bytes32 rootAfter, uint64 noteAmount);
+    event NewNote(bytes32 indexed commitment, uint32 leafIndex, bytes32 rootAfter, uint128 noteAmount);
 
     /// @notice A note was spent via mint.
     /// @param noteOwner Owner proven by the mint proof (indexed).
@@ -45,6 +46,6 @@ interface IEmit {
     /// @param nullifier The spent nullifier (indexed).
     /// @param mintAmount Credited native base units.
     event NoteUsed(
-        address indexed noteOwner, address indexed payoutRecipient, bytes32 indexed nullifier, uint64 mintAmount
+        address indexed noteOwner, address indexed payoutRecipient, bytes32 indexed nullifier, uint128 mintAmount
     );
 }
